spider/requirements.txt
Django==1.11.3
django-filter==2.0.0
djangorestframework==3.8.2
pytz==2018.5
